(mostly) use Shortcut rather than Keys
Thus it's not required for the root Presentation to retain key focus.
That way each slide can also use the keyboard as long as those uses
don't interefere with the global shortcuts.  The shortcuts can also be
disabled in case they interfere (e.g. using arrow navigation globally
prevents using it for navigating within a slide).

The remaining exception is for directly entering slide numbers.
IMO this is impractical anyway; but if you need to do that, then
you also need to give the Presentation keyboard focus, as before.

The shortcut which used Esc to quit is removed, and replaced with
the more-standard Control-Q.  Enter and Backspace are also no longer
used for slide navigation: there seem to be enough keys for that.

Change-Id: I0a54ae62305aa28b0ffef15b5ec010058c36fb8e
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>